Reorganize section_break interface
