+++ app/core/UIManager.js
@@ -12,7 +12,7 @@ define(function(require, exports, module) {
     require('core/interfaces/_internals/views/interface'),
     require('core/interfaces/_internals/file_preview/interface'),
     require('core/interfaces/textinput/index'),
-    require('core/interfaces/section_break/section_break'),
+    require('core/interfaces/section_break/component'),
     require('core/interfaces/checkbox/checkbox'),
     require('core/interfaces/color/index'),
     require('core/interfaces/numeric/index'),
